Refractor API into folders. Work on API functions for Places and Location
