+++ src/api/index.ts
@@ -1,8 +1,8 @@
-import * as Auth from "./auth.api";
-// import * as Location from "./location.api";
-import * as Token from "./token.api";
-import * as User from "./user.api";
-import * as Place from "./place.api";
+import * as Auth from "./auth";
+// import * as Location from "./location";
+import * as Token from "./auth/token";
+import * as User from "./users";
+import * as Place from "./places";
 import * as HereMaps from "./heremaps.api";
 
 export const BannerAPI = {
