+++ src/api/location/index.ts
@@ -1,0 +1,7 @@
+export const getPlacesNear = async (): Promise<void> => {
+  return;
+};
+
+export const getPlaceInfoNear = async (): Promise<void> => {
+  return;
+};
