+++ src/api/places/index.ts
@@ -1,0 +1,8 @@
+import { Place } from "../../types/Place";
+import * as Location from "expo-location";
+
+export const getPlaceInfo = async (id: string): Promise<void> => {};
+
+export const getPlaceByAddress = async (): Promise<void> => {
+  return;
+};
